services/identity/pyproject.toml
[project]
name = "onchange-identity"
version = "0.1.0"
description = "Identity service for onChange by Sairo — orgs, users, memberships, sessions, API keys."
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Sairo" }]
readme = "README.md"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "argon2-cffi>=23.1",
    "python-jose[cryptography]>=3.3",
    "uuid6>=2024.7.10",
    "structlog>=24.4",
    "orjson>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-python-jose",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "PL"]
ignore = ["PLR0913"]  # allow many-args for FastAPI dependency-injection

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
